api/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pdfplumber>=0.10.3
langchain>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0,<3.0.0
pinecone-client>=5.0.0
langchain-pinecone>=0.0.1
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
